requirements.txt
langchain-core==1.3.2
langchain-anthropic==1.4.2
python-telegram-bot==22.7
python-dotenv==1.2.2
